% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{detection_models}
\alias{detection_models}
\title{Occupancy modeling that produces detection probabilities}
\format{
A named list containing 30 models (10) species x 3 survey durations)
}
\source{
Alberta Biodiversity Monitoring Institute
}
\usage{
detection_models
}
\description{
Occupancy modeling that produces detection probabilities
}
\keyword{data}
