% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.delete_wav_files}
\alias{.delete_wav_files}
\title{Internal function to delete media}
\usage{
.delete_wav_files(dir)
}
\arguments{
\item{dir}{Directory containing files}
}
\description{
Internal function to delete media.
}
\keyword{internal}
