% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.language}
\alias{.language}
\title{Switch locale to another language}
\usage{
.language(language = c("en", "fr"))
}
\description{
Global function to allow a user to request data in another language. Currently English = en or French = fr.
}
\keyword{internal}
