% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_qpad.R
\name{.make_off}
\alias{.make_off}
\title{Design matrices for singing rates (\code{Xp}) and for EDR (\code{Xq})}
\usage{
.make_off(spp, x, useMethod = "y")
}
\description{
Design matrices for singing rates (\code{Xp}) and for EDR (\code{Xq})
}
