% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_qpad.R
\name{.make_x}
\alias{.make_x}
\title{Internal functions}
\usage{
.make_x(data, tz = "local", check_xy = TRUE)
}
\arguments{
\item{data}{Dataframe output from the \code{wt_format_wide} function.}

\item{tz}{Character; whether or not the data is in local or UTC time ("local", or "utc"). Defaults to "local".}

\item{check_xy}{Logical; check whether coordinates are within the range that QPAD offsets are valid for.}
}
\description{
Functions to format reports for qpad offset calculation.
}
\details{
QPAD offsets, wrapped by the \code{wt_qpad_offsets} function.
}
