% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.wt_api_put}
\alias{.wt_api_put}
\title{An internal function to handle generic PUT requests to WildTrax
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}}
\usage{
.wt_api_put(path, body = list(), max_time = 300)
}
\arguments{
\item{path}{The path to the API}

\item{body}{A named list to include in the request body (JSON)}

\item{max_time}{The maximum number of seconds an API request can take. By default 300.}
}
\description{
Generic function to handle certain PUT requests
}
\keyword{internal}
