% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.wt_auth}
\alias{.wt_auth}
\title{Internal functions}
\usage{
.wt_auth()
}
\description{
Get Auth0 token and assign information to the hidden environment
}
\details{
WildTrax authentication
}
\keyword{internal}
