% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.wt_calculate_prf}
\alias{.wt_calculate_prf}
\title{Internal evaluation function for acoustic classifiers}
\usage{
.wt_calculate_prf(threshold, data, human_total)
}
\arguments{
\item{threshold}{A single numeric value for score threshold}

\item{data}{Output from the \code{wt_download_report()} function when you request the \code{main} and \code{birdnet} reports}

\item{human_total}{The total number of detections in the gold standard, typically from human listening data (e.g., the main report)}
}
\value{
A vector of precision, recall, F-score, and threshold
}
\description{
Internal function to calculate precision, recall, and F-score for a given score threshold.
}
\keyword{internal}
