% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wt_prob_det.R
\name{run_prob_det_api}
\alias{run_prob_det_api}
\title{Run prob_det function API}
\usage{
run_prob_det_api(...)
}
\arguments{
\item{...}{Options passed to \code{plumber::plumb()$run()}}
}
\value{
A running Plumber API
}
\description{
Run prob_det function API
}
