% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wildRtrax-package.R
\docType{package}
\name{wildRtrax-package}
\alias{wildRtrax}
\alias{wildRtrax-package}
\title{wildRtrax: Environmental sensor data management and analytics to and from WildTrax}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

wildRtrax is an R package that enables environmental sensor data users to create comprehensive work flows for managing and analyzing data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://abbiodiversity.github.io/wildRtrax/}
}

}
\author{
\strong{Maintainer}: Marcus Becker \email{mabecker@ualberta.ca}

Authors:
\itemize{
  \item Alex MacPhail \email{agmacpha@ualberta.ca}
  \item Elly Knight \email{ecknight@ualberta.ca}
}

}
\keyword{internal}
