% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wildRtrax.R
\docType{package}
\name{wildrtrax}
\alias{wildrtrax-package}
\alias{wildrtrax}
\title{wildrtrax: A series of functions for environmental sensor data management and analysis to and from WildTrax}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

This R package contains functions to help manage and analyze environmental sensor data to and from WildTrax
}
\seealso{
Useful links:
\itemize{
  \item \url{https://abbiodiversity.github.io/wildrtrax/}
}

}
\author{
Alex MacPhail \email{agmacpha@ualberta.ca}
}
