% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience-functions.R
\name{wt_add_grts}
\alias{wt_add_grts}
\title{Intersect locations to add a GRTS ID}
\usage{
wt_add_grts(data, group_locations_in_cell = FALSE)
}
\arguments{
\item{data}{Data containing locations}

\item{group_locations_in_cell}{Option to provide distinct location names if points are found in the same cell. Sequentially provides a number for each GRTS ID e.g. 3-1, 3-2, etc.}
}
\value{
A dataframe with the additional GRTS IDs
}
\description{
This function intersects location data with the GRTS ID provided by \href{https://www.nabatmonitoring.org/}{NABat}
}
\examples{
\dontrun{

dat.grts <- wt_download_report(reports = "location")
grts.data <- wt_add_grts(dat.grts)
}
}
