% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acoustic-pre-processing.R
\name{wt_audio_scanner}
\alias{wt_audio_scanner}
\title{Scan acoustic data to a standard format}
\usage{
wt_audio_scanner(path, file_type, extra_cols = FALSE)
}
\arguments{
\item{path}{Character; The path to the directory with audio files}

\item{file_type}{Character; File types of wav, wac, flac or all}

\item{extra_cols}{Boolean; TRUE returns duration, sample rate and channels}
}
\value{
A tibble with a summary of your audio files.
}
\description{
Scans directories of audio data
}
\examples{
\dontrun{
wt_audio_scanner(path = ".", file_type = "wav", extra_cols = T)
}

}
