% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{wt_auth}
\alias{wt_auth}
\title{Authenticate into WildTrax}
\usage{
wt_auth(force = FALSE)
}
\arguments{
\item{force}{Logical; whether or not the force re-authentication even if token has not expired. Defaults to FALSE.}
}
\value{

}
\description{
Obtain Auth0 credentials using WT_USERNAME and WT_PASSWORD stored as environment variables
}
\examples{
\dontrun{
# Authenticate first:
wt_auth(force = FALSE)
}

}
