% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wt_cam_report_tag_summary.R
\name{wt_cam_report_tag_summary}
\alias{wt_cam_report_tag_summary}
\title{Retrieve tagged image data from camera project}
\usage{
wt_cam_report_tag_summary(conn, project_id, native_only = TRUE)
}
\arguments{
\item{conn}{WildTrax database connection (PostgreSQLConnection object)}

\item{project_id}{Integer; the project ID}

\item{native_only}{Logical; set to TRUE to retrieve tagged data on native species}
}
\value{
This function returns a dataframe of tagged camera image data from the project specified in project_id
}
\description{
Retrieve tagged image data from camera project
}
