% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/camera-pre-processing.R
\name{wt_check_datetime_constant}
\alias{wt_check_datetime_constant}
\title{Identify groups with constant image datetimes}
\usage{
wt_check_datetime_constant(x)
}
\description{
Internal helper function used by \code{\link[=wt_image_check]{wt_image_check()}} to identify groups where
all images have the same datetime, which may indicate an incorrectly set
camera clock.
}
\keyword{internal}
