% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/camera-pre-processing.R
\name{wt_check_datetime_jan1}
\alias{wt_check_datetime_jan1}
\title{Identify image datetimes occurring on January 1}
\usage{
wt_check_datetime_jan1(x)
}
\description{
Internal helper function used by \code{\link[=wt_image_check]{wt_image_check()}} to identify groups
containing image timestamps on January 1, which can indicate incorrectly
initialized camera clocks.
}
\keyword{internal}
