% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/camera-pre-processing.R
\name{wt_check_datetime_month}
\alias{wt_check_datetime_month}
\title{Check image datetimes against expected months}
\usage{
wt_check_datetime_month(x, months = NULL)
}
\description{
Internal helper function used by \code{\link[=wt_image_check]{wt_image_check()}} to identify groups
containing image datetimes outside the expected months.
}
\keyword{internal}
