% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/camera-pre-processing.R
\name{wt_check_datetime_year}
\alias{wt_check_datetime_year}
\title{Check image datetimes against expected years}
\usage{
wt_check_datetime_year(x, years = NULL)
}
\description{
Internal helper function used by \code{\link[=wt_image_check]{wt_image_check()}} to identify groups
containing image datetimes outside the expected years.
}
\keyword{internal}
