% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/camera-pre-processing.R
\name{wt_check_invalid_datetime}
\alias{wt_check_invalid_datetime}
\title{Identify invalid image datetimes}
\usage{
wt_check_invalid_datetime(x, min_year = 1995)
}
\arguments{
\item{min_year}{Minimum acceptable year for image timestamps.}
}
\description{
Internal helper function used by \code{\link[=wt_image_check]{wt_image_check()}} to identify groups with
image timestamps occurring before a minimum year or in the future.
}
\keyword{internal}
