% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifier-functions.R
\name{wt_classifier_detections}
\alias{wt_classifier_detections}
\title{Convert to detections}
\usage{
wt_classifier_detections(
  data,
  threshold,
  remove_species = TRUE,
  species = NULL
)
}
\arguments{
\item{data}{The \code{birdnet} report from the \code{wt_download_report()} function}

\item{threshold}{Numeric; the desired score threshold}

\item{remove_species}{Logical; indicates whether species that are not allowed in the WildTrax project should be removed from the BirdNET report}

\item{species}{Character; optional subset of species to calculate metrics for (e.g., species = c("OVEN", "OSLF", "BOCH"))}
}
\value{
A tibble with the same fields as the \code{birdnet} report that contains only detections above the specified threshold.
}
\description{
Converts and filters the \code{birdnet} report from score probabilities to detections using a specified score threshold. You can exclude species that are not allowed in the project from the BirdNET results before evaluation.
}
\examples{
\dontrun{
data <- wt_download_report(project_id = 1144, sensor_id = "ARU", reports = c("main", "birdnet"), weather_cols = FALSE)

eval <- wt_evaluate_classifier(data, resolution = "recording", remove_species = TRUE, thresholds = c(10, 99))

threshold_use <- wt_get_threshold(eval) |> print()

birdnet <- data[[1]]

detections <- wt_classifier_detections(birdnet, threshold = threshold_use, remove_species = TRUE)
}

}
