% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wt_collect_cam_meta.R
\name{wt_collect_cam_meta}
\alias{wt_collect_cam_meta}
\title{Retrieve camera deployment metadata}
\usage{
wt_collect_cam_meta(con, year, project)
}
\arguments{
\item{con}{PostgreSQLConnection object to Wildtrax}

\item{year}{Character string to filter by year}

\item{project}{Character string of full project name to filter by project}
}
\value{
This function returns a \code{data.frame} of the metadata associated with the camera deployments requested.
}
\description{
Retrieve camera deployment metadata
}
\author{
Marcus Becker

Alex MacPhail
}
