% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wt_collect_table.R
\name{wt_collect_table}
\alias{wt_collect_table}
\title{Collect a table from the Wildtrax PostgreSQL database}
\usage{
wt_collect_table(con, schema, table)
}
\arguments{
\item{con}{PostgreSQLConnection object to Wildtrax}

\item{schema}{Character string indicating the schema that the table is in}

\item{table}{Character string indicating the table}
}
\value{
This function returns a \code{data.frame} of the table requested.
}
\description{
Collect a table from the Wildtrax PostgreSQL database
}
\author{
Marcus Becker

Alex MacPhail
}
