% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wt_download_clips.R
\name{wt_download_clips}
\alias{wt_download_clips}
\title{Download clips of WildTrax annotations.}
\usage{
wt_download_clips(data, filepath, object = "soundfile", quiet = FALSE)
}
\arguments{
\item{data}{Dataframe of annotations obtained from the wt_download_tagreport function.}

\item{filepath}{Folder to download objects into.}

\item{object}{The format requested for the clips; either "soundfile", or "spectrogram".}

\item{quiet}{Suppress download details for each object.}
}
\value{
A dataframe identical to input but including ID and filepath for each downloaded object.
}
\description{
This function downloads annotations from WildTrax either as sound files or spectrograms. Using this function requires obtaining the urls for the objects using the wt_download_tagreport function.
}
\examples{
\dontrun{
dat.tag <- wt_download_clips(dat, "data/clips", "soundfile", quiet=TRUE)
}
}
