% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{wt_download_report}
\alias{wt_download_report}
\title{Download formatted reports from WildTrax}
\usage{
wt_download_report(project_id, sensor_id, reports, max_seconds = 300)
}
\arguments{
\item{project_id}{Numeric; the project ID number that you would like to download data for. Use \code{wt_get_projects()} to retrieve these IDs.}

\item{sensor_id}{Character; Can either be "ARU", "CAM", or "PC".}

\item{reports}{Character; The report type to be returned. Multiple values are accepted as a concatenated string.}

\item{max_seconds}{Numeric; Number of seconds to force to wait for downloads.}
}
\value{
If multiple report types are requested, a list object is returned; if only one, a dataframe.
}
\description{
Download various ARU, camera, or point count data from projects across WildTrax
}
\details{
Valid values for argument \code{report} when \code{sensor_id} = "CAM" currently are:
\itemize{
\item main
\item project
\item location
\item image_report
\item image_set_report
\item tag
\item megadetector
\item definitions
}

Valid values for argument \code{report} when \code{sensor_id} = "ARU" currently are:
\itemize{
\item main
\item project
\item location
\item recording
\item tag
\item ai
\item definitions
}

Valid values for argument \code{report} when \code{sensor_id} = "PC" currently are:
\itemize{
\item main
\item project
\item location
\item point_count
\item definitions
}
}
\examples{
\dontrun{
# Authenticate first:
wt_auth()
a_camera_project <- wt_download_report(
project_id = 205, sensor_id = "CAM", reports = c("tag", "image_set_report"))

an_aru_project <- wt_download_report(
project_id = 3261, sensor_id = "ARU", reports = c("main", "ai"))
}

}
