% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wt_download_tagreport.R
\name{wt_download_tagreport}
\alias{wt_download_tagreport}
\title{Download report of tags.}
\usage{
wt_download_tagreport(project_id)
}
\arguments{
\item{project_id}{Numeric; the project ID number that you would like to download data for. Use \code{wt_get_download_summary()} to retrieve these IDs.}
}
\value{
Dataframe of annotations and associated metadata.
}
\description{
This function downloads a report of all tags or annotations for an ARU project. The report contains urls for downloading clips of those annotations. This function should be used prior to using the \code{wt_download_clips()} function.
}
\examples{
\dontrun{
# Authenticate first:
wt_auth()

tags <- wt_download_tagreport(project_id = 397)
}

}
