% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{wt_download_tags}
\alias{wt_download_tags}
\title{Download acoustic tags}
\usage{
wt_download_tags(input, output, clip_type = c("spectrogram", "audio"))
}
\arguments{
\item{input}{A data frame or tibble of the tag report i.e. wt_download_report(reports = "tag")}

\item{output}{Directory to store the tags}

\item{clip_type}{Character; either spectrogram or audio clips}
}
\value{
An organized folder of clips and tags in the output directory. Assigning wt_download_tags to an object will return the table form of the data with the functions returning the after effects in the output directory
}
\description{
Downloads acoustic tags (mp3 and jpg) in an organized format for other purposes.
}
\examples{
\dontrun{
dat.report <- wt_download_report(reports = "tag")
wt_download_tags(input = my_tag_data)
}

}
