% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acoustic-pre-processing.R
\name{wt_flac_info}
\alias{wt_flac_info}
\title{Extract relevant metadata from a flac file}
\usage{
wt_flac_info(path)
}
\arguments{
\item{path}{Character; The flac file path}
}
\value{
a list with relevant information
}
\description{
Scrape relevant information from flac file
}
