% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience-functions.R
\name{wt_format_data}
\alias{wt_format_data}
\title{Format data for a specified portal}
\usage{
wt_format_data(input, format = "FWMIS")
}
\arguments{
\item{input}{The report from \code{wt_download_report()}}

\item{format}{A format i.e. 'FWMIS'}
}
\value{
A tibble with the formatted report
}
\description{
This function takes the WildTrax reports and converts them to the desired format
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
\dontrun{

dat <- wt_download_report(reports = c("main","visit","equipment")) |>
wt_format_data(format = 'FWMIS')
}
}
