% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience-functions.R
\name{wt_format_occupancy}
\alias{wt_format_occupancy}
\title{Format WildTrax report for occupancy modelling}
\usage{
wt_format_occupancy(data, species, siteCovs = NULL)
}
\arguments{
\item{data}{Summary report of WildTrax observations from the \code{wt_download_report()} function. Currently only functioning for the ARU sensor.}

\item{species}{Character; four-letter alpha code for the species desired for occupancy modelling.}

\item{siteCovs}{Optional dataframe of site covariates. Must contain a column with the same values as the location field in the data, with one row per unique value of location (i.e., one row per site).}
}
\value{
An object of class unmarkedFrameOccu. See \code{?unmarked::unmarkedFrameOccu} for details.
}
\description{
This function formats the summary report from the \code{wt_download_report()} function into an unmarked object for occupancy modelling. The current version only includes formatting for the ARU sensor and for single species single season models.
}
\examples{
\dontrun{
dat.occu <- wt_format_occupancy(dat, species="CONI", siteCovs=NULL)
mod <- occu(~ 1 ~ 1, dat.occu)
}
}
