% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{wt_get_download_summary}
\alias{wt_get_download_summary}
\title{Get download summary}
\usage{
wt_get_download_summary(sensor_id)
}
\arguments{
\item{sensor_id}{Can either be "ARU" or "CAM"}
}
\value{
a dataframe listing the projects that the user can download data for, including: project name, id, year, number of tags, and status.
}
\description{
Obtain a table listing projects that the user is able to download data for
}
\examples{
\dontrun{
# Authenticate first:
wt_auth()
wt_get_download_summary(sensor_id = "ARU")
}

}
