% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{wt_get_image_sets}
\alias{wt_get_image_sets}
\title{Get an Organizations' list of image sets}
\usage{
wt_get_image_sets(organization)
}
\arguments{
\item{organization}{Either the short letter or numeric digit representing the Organization}
}
\value{
A data frame listing an Organizations' recordings
}
\description{
Obtain a table listing all image sets belonging to an Organization emulating the Image Sets tab
}
\examples{
\dontrun{
# Authenticate first:
wt_auth()
wt_get_image_sets(organization = 'ABMI')
}

}
