% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{wt_get_organizations}
\alias{wt_get_organizations}
\title{Get a summary of available Organizations from WildTrax}
\usage{
wt_get_organizations()
}
\value{
A data frame listing the Organization that the user can access including important Organization summaries
}
\description{
Obtain a table listing Organizations that the user is able to download data for.
}
\examples{
\dontrun{
# Authenticate first:
wt_auth()
wt_get_organizations()
}

}
