% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{wt_get_projects}
\alias{wt_get_projects}
\title{Get a summary of available projects from WildTrax}
\usage{
wt_get_projects(sensor)
}
\arguments{
\item{sensor}{Can be one of "ARU", "CAM", or "PC"}
}
\value{
A data frame listing the projects that the user can download data for, including: project name, id, year, number of tasks and project status.
}
\description{
Obtain a table listing projects that the user is able to download data for.
}
\examples{
\dontrun{
# Authenticate first:
wt_auth()
wt_get_projects(sensor = "ARU")
}

}
