% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{wt_get_recordings}
\alias{wt_get_recordings}
\title{Get an Organizations' list of recordings}
\usage{
wt_get_recordings(organization)
}
\arguments{
\item{organization}{Either the short letter or numeric digit representing the Organization}
}
\value{
A data frame listing an Organizations' recordings
}
\description{
Obtain a table listing all recordings belonging to an Organization emulating the Recordings tab
}
\examples{
\dontrun{
# Authenticate first:
wt_auth()
wt_get_recordings(organization = 'ABMI')
}

}
