% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{wt_get_species}
\alias{wt_get_species}
\title{Get the WildTrax species table}
\usage{
wt_get_species(get_all = NULL)
}
\arguments{
\item{get_all}{Don't truncate to important columns - gimme the whole species table!}
}
\value{
A tibble of the WildTrax species table
}
\description{
Request for the WildTrax species table
}
\examples{
 \dontrun{
 wt_species <- wt_get_species()
 }
}
