% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{wt_get_species_presets}
\alias{wt_get_species_presets}
\title{Get the WildTrax species preset lists}
\usage{
wt_get_species_presets(preset = NULL)
}
\arguments{
\item{preset}{The name of the preset list}
}
\value{
A tibble of the WildTrax species list for a specific project
}
\description{
Request for the WildTrax species list for a project
}
\examples{
 \dontrun{
 presets <- wt_get_species_presets()
 }
}
