% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{wt_get_view}
\alias{wt_get_view}
\title{Get data from WildTrax views}
\usage{
wt_get_view(
  api,
  project = NULL,
  organization = NULL,
  max_seconds = 300,
  ai_threshold = 0.5
)
}
\arguments{
\item{api}{A string specifying the API to query. Must be one of:
\itemize{
\item \code{"organization_locations"}
\item \code{"organization_visits"}
\item \verb{"organization_equipment}"
\item \verb{"organization_deployments}"
\item \verb{"organization_recordings}"
\item \verb{"organization_image_sets}"
\item \verb{"organization_usage_report}
\item \code{"project_aru_tasks"}
\item \code{"project_camera_tasks"}
\item \code{"project_point_counts"}
}}

\item{project}{Numeric; The project id}

\item{organization}{Numeric; The organization id}

\item{max_seconds}{Numeric; Number of seconds to force to wait for downloads}

\item{ai_threshold}{Numeric; Minimum threshold set for AI classifier detections}
}
\value{
A tibble with column headers for the specified API call.
}
\description{
Fetch data for tables and views in WildTrax. You must specify at least one of \code{project} or \code{organization} depending on the API
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
\dontrun{
# Authenticate first:
wt_auth()

# Fetch locations by organization
wt_get_view("organization_locations", organization = 5205)

# Fetch locations by project
wt_get_view("project_locations", project = 620)
}

}
