% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{wt_get_views}
\alias{wt_get_views}
\title{Get data from WildTrax views}
\usage{
wt_get_views(api, project = NULL, organization = NULL)
}
\arguments{
\item{api}{A string specifying the API to query. Must be one of:
\itemize{
\item \code{"organization_locations"}
\item \code{"organization_visits"}
\item \verb{"organization_equipment}"
\item \verb{"organization_deployments}"
\item \verb{"organization_recordings}"
\item \code{"project_locations"}
\item \code{"project_species"}
\item \code{"project_aru_tasks"}
\item \code{"project_aru_tags"}
\item \code{"project_image_metadata"}
\item \code{"project_camera_tags"}
\item \code{"project_point_counts"}
}}

\item{project}{The project id}

\item{organization}{The organization id}
}
\value{
A tibble with column headers for the specified API call.
}
\description{
Fetch data for tables and views in WildTrax. You must specify at least one of \code{project} or \code{organization} depending on the API
}
\examples{
\dontrun{
# Authenticate first:
wt_auth()

# Fetch locations by organization
wt_get_views("organization_locations", organization = "ABMI")

# Fetch locations by project
wt_get_views("project_locations", project = 620)
}

}
