% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{wt_get_visits}
\alias{wt_get_visits}
\title{Get visits from a WildTrax Organization}
\usage{
wt_get_visits(organization)
}
\arguments{
\item{organization}{Either the short letter or numeric digit representing the Organization}

\item{location}{Supply a location to get all visits from that location}
}
\value{
A data frame listing an Organizations' visits
}
\description{
Obtain a table listing visits emulating the Visits tab in a WildTrax Organization
}
\examples{
\dontrun{
# Authenticate first:
wt_auth()
wt_get_visits(organization = 'ABMI')
}

}
