% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acoustic-pre-processing.R
\name{wt_glean_ap}
\alias{wt_glean_ap}
\title{Extract and plot relevant acoustic index metadata and LDFCs}
\usage{
wt_glean_ap(x = NULL, input_dir, purpose = c("quality", "abiotic", "biotic"))
}
\arguments{
\item{x}{A data frame or tibble; must contain the file name. Use output from \code{`wt_audio_scanner`}.}

\item{input_dir}{Character; A folder path where outputs from \code{`wt_run_ap`} are stored.}

\item{purpose}{Character; type of filtering you can choose from}
}
\value{
Output will return the merged tibble with all information, the summary plots of the indices and the LDFC
}
\description{
This function will use a list of media files from a \verb{wt_*} work flow and outputs from \code{wt_run_ap}
in order to generate summary plots of acoustic indices and long-duration false-colour spectrograms. This can
be viewed as the "final step" in interpreting acoustic index and LDFC values from your recordings.
}
\examples{
\dontrun{
wt_glean_ap(x = wt_audio_scanner_data, input_dir = "/path/to/my/files")
}

}
