% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acoustic-pre-processing.R
\name{wt_guano_tags}
\alias{wt_guano_tags}
\title{Convert GUANO embeded metadata to tags and metadata output for a Project}
\usage{
wt_guano_tags(path, output = FALSE, output_file = NULL)
}
\arguments{
\item{path}{Character; The path to the input csv}

\item{output}{Character; Path where the output file will be stored}

\item{output_file}{Character; Path of the output file}
}
\value{
A csv formatted as a WildTrax tag template
}
\description{
\code{wt_guano_tags} Takes the embeded classifier output and converts them into a WildTrax tag template for upload
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
\dontrun{
# Process a single audio file
wt_guano_tags("/path/to/audio_file.wav")

# Process audio files from a directory
wt_audio_scanner("/path/to/audio", file_type = "wav", extra_cols = TRUE) |>
  purrr::map(\(x) = file_path ~wt_guano_tags(file_path)) |>
  bind_rows()
}

}
