% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze-data.R
\name{wt_ind_detect}
\alias{wt_ind_detect}
\title{Evaluate independent detections in your camera data}
\usage{
wt_ind_detect(
  x,
  threshold,
  units = "minutes",
  datetime_col = image_date_time,
  remove_human = TRUE,
  remove_domestic = TRUE
)
}
\arguments{
\item{x}{A dataframe of camera data; preferably, the main report from \code{wt_download_report()}.}

\item{threshold}{Numeric; time interval to parse out independent detections.}

\item{units}{The threshold unit. Can be one of three values, "seconds", "minutes", "hours".}

\item{datetime_col}{Defaults to \code{image_date_time}; The column indicating the timestamp of the image.}

\item{remove_human}{Logical; Should human and human-related tags (e.g. vehicles) be removed? Defaults to TRUE.}

\item{remove_domestic}{Logical; Should domestic animal tags (e.g. cows) be removed? Defaults to TRUE.}
}
\value{
A dataframe of independent detections in your camera data, based on the threshold you specified. The df wil include information about the duration of each detection, the number of images, the average number of individual animals per image, and the max number of animals in the detection.
}
\description{
Create independent detections dataframe using camera data from WildTrax
}
\examples{
\dontrun{
detections <- wt_ind_detect(x = df, threshold = 30, units = "minutes")
}

}
