% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acoustic-pre-processing.R
\name{wt_kaleidoscope_tags}
\alias{wt_kaleidoscope_tags}
\title{Convert Kaleidoscope output to tags}
\usage{
wt_kaleidoscope_tags(input, output, freq_bump = T)
}
\arguments{
\item{input}{Character; The path to the input csv}

\item{output}{Character; Path where the output file will be stored}

\item{freq_bump}{Boolean; Set to TRUE to add a buffer to the frequency values exported from Kaleidoscope. Helpful for getting more context around a signal in species verification}
}
\value{
A csv formatted as a WildTrax tag template
}
\description{
\code{wt_kaleidoscope_tags} Takes the classifier output from Wildlife Acoustics Kaleidoscope and converts them into a WildTrax tag template for upload
}
\examples{
\dontrun{
wt_kaleidoscope_tags(input = input.csv, output = tags.csv, freq_bump = T)
}

}
