% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{wt_location_photos}
\alias{wt_location_photos}
\title{Batch upload and download locations photos}
\usage{
wt_location_photos(data, direction, dir)
}
\arguments{
\item{data}{When going up, provide the joining data for locations and visits}

\item{direction}{Either "up" or "down"; if "up" need to supply the appropriate data to join to a location and / or visit}

\item{dir}{Folder of images either going "up" or "down"}
}
\value{
When "down", a folder of images
}
\description{
A two-way street to upload and download location photos
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
\dontrun{
# Authenticate first:
wt_auth()
# When going up provide the folder of images and data for joining
wt_location_photos(organization = 'ABMI', data = "/my/join/table",
direction = "up", dir = "/my/dir/of/images")

# When going down the directory where you want the files only
wt_location_photos(organization = 'ABMI', direction = "down",
dir = "/my/dir/to/download/images")

}

}
