% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze-data.R
\name{wt_ord}
\alias{wt_ord}
\title{Use a series of multivariate analyses to determine observer differences in acoustic processed data}
\usage{
wt_ord(input = x, min_obs, confidence)
}
\arguments{
\item{input}{Character; A wt_download_report tibble}

\item{min_obs}{Numeric; The minimum number of replicates you want to use. wt_ord will omit}

\item{confidence}{Numeric; The confidence of the ellipses in the RDA}
}
\value{
A list containing the following:
}
\description{
Uses a series of multivariate analyses to determine differences in multiobserver data
}
\section{\code{wt_ord} details}{

}

\examples{
\dontrun{
res <- wt_prd(input = data, min_obs = 10, confidence = 0.67)
}

}
