% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/camera-pre-processing.R
\name{wt_print_datetime_report}
\alias{wt_print_datetime_report}
\title{Print image datetime QA/QC results}
\usage{
wt_print_datetime_report(results, verbose = FALSE)
}
\description{
Internal helper function used by \code{\link[=wt_image_check]{wt_image_check()}} to format and print
datetime quality control results.
}
\keyword{internal}
