% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wt_prob_det.R
\name{wt_prob_det}
\alias{wt_prob_det}
\title{Estimate the probability of detection, given a location is occupied, based on the audio survey length and the number of surveys conducted.}
\usage{
wt_prob_det(species_code, survey_length, number_of_surveys)
}
\arguments{
\item{species_code}{Character; the species code. See below for included species.}

\item{survey_length}{Double; the survey length in minutes. Valid values are 1, 2, and 3.}

\item{number_of_surveys}{Double; the number of surveys (to be) conducted.}
}
\value{
Double; the estimated probability of species detection based on the number of surveys planned.
}
\description{
Using data from the ABMI ecosystem health acoustic surveys (2015-2020), occupancy models were fit to five common boreal songbirds. These models are used to estimate the detection probability for each species for surveys of different lengths (1-3 minutes).
Given that it is present, these results can be used to estimate the probability the species will be detected at least once as a function of the number of surveys conducted.
}
\details{
Valid values for species_code currently include:
\itemize{
\item "OVEN" (Ovenbird)
\item "CCSP" (Clay-colored Sparrow)
\item "OSFL" (Olive-sided Flycatcher)
\item "TEWA" (Tennessee Warbler)
\item "WTSP" (White-throated Sparrow)
\item "ALFL" (Alder Flycatcher)
\item "LISP" (Lincolns Sparrow)
\item "YRWA" (Yellow-rumped Warbler)
\item "CONI" (Common Nighthawk)
\item "SWTH" (Swainsons Thrush)
\item "WOLF" (Gray Wolf)
\item "GGOW" (Great Gray Owl)
\item "GHOW" (Great Horned Owl)
\item "NSWO" (Northern Saw-whet Owl)
\item "SEOW" (Short-eared Owl)
}
}
\examples{
\dontrun{
wt_prob_det(species_code = "OVEN", survey_length = 1, number_of_surveys = 4)
}

}
