% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acoustic-pre-processing.R
\name{wt_run_ap}
\alias{wt_run_ap}
\title{Get a variety of acoustic index output from audio}
\usage{
wt_run_ap(
  x = NULL,
  fp_col = file_path,
  audio_dir = NULL,
  output_dir,
  path_to_ap = "C:\\\\AP\\\\AnalysisPrograms.exe"
)
}
\arguments{
\item{x}{(optional) A data frame or tibble; must contain the absolute audio file path and file name. Use output from \code{`wt_audio_scanner`}.}

\item{fp_col}{If x is supplied, the column containing the audio file paths. Defaults to file_path.}

\item{audio_dir}{(optional) Character; path to directory storing audio files.}

\item{output_dir}{Character; path to directory where you want outputs to be stored.}

\item{path_to_ap}{Character; file path to the AnalysisPrograms software package. Defaults to "C:\\AP\\AnalysisPrograms.exe".}
}
\value{
Output will return to the specific root directory
}
\description{
For generating acoustic indices and false-colour spectrograms using QUT Ecoacoustics \strong{A}nalysis \strong{P}rograms software. See \url{https://github.com/QutEcoacoustics/audio-analysis} for information about usage and installation of the AP software.
Note that this function relies on having this software installed locally.

This function will batch calculate summary and spectral acoustic indices and generate false-colour spectrograms for a folder of audio files using the Towsey.Acoustic configuration (yml) file from the AP software.
You can use the output from \code{`wt_audio_scanner`} in the function, or define a local folder with audio files directly.
}
