% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wt_prob_det.R
\name{wt_run_shiny}
\alias{wt_run_shiny}
\title{Run a wildRtrax Shiny App (locally)}
\usage{
wt_run_shiny(app)
}
\arguments{
\item{app}{Character; name of the shiny app.}
}
\value{
A shiny app running locally
}
\description{
Choose a shiny app included as part of the wildRtrax package and run it locally
}
