% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wt_signal_level.R
\name{wt_signal_level}
\alias{wt_signal_level}
\title{Extract relative sound level from a wav file}
\usage{
wt_signal_level(
  path,
  fmin = 500,
  fmax = NA,
  threshold,
  channel = "left",
  aggregate = NULL
)
}
\arguments{
\item{path}{The path to the wav file}

\item{fmin}{The frequency minimum}

\item{fmax}{The frequency maximum}

\item{threshold}{The desired threshold}

\item{channel}{Choose "left" or "right"}

\item{aggregate}{Aggregate detections by this number of seconds, if desired}
}
\value{
A list object containing the following four elements: output (dataframe), aggregated (boolean), channel (character), and threshold (numeric)
}
\description{
Extract relative sound level from a wav file
}
\examples{
\dontrun{
df <- wt_signal_level(path = "")
}

}
