% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acoustic-pre-processing.R
\name{wt_songscope_tags}
\alias{wt_songscope_tags}
\title{Convert Songscope output to tags}
\usage{
wt_songscope_tags(
  input,
  output = c("env", "csv"),
  output_file = NULL,
  species,
  vocalization,
  score_filter,
  method = NULL,
  duration,
  sample_freq
)
}
\arguments{
\item{input}{Character; The path to the input csv}

\item{output}{Character; Path where the output file will be stored}

\item{output_file}{Character; Path of the output file}

\item{species}{Character; Short-hand code for the species (see \code{wt_get_species()})}

\item{vocalization}{Character; The vocalization type from either Song, Call, Non-Vocal, Nocturnal flight call and Feeding Buzz}

\item{score_filter}{Numeric; Filter the detections by a score threshold}

\item{method}{Character; Include options from 1SPT, 1SPM or None}

\item{duration}{Numeric; length of the task in seconds}

\item{sample_freq}{Numeric; The sampling frequency in Hz of the recording}
}
\value{
A csv formatted as a WildTrax tag template
}
\description{
Convert Songscope output to tags
}
