% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.make_off}
\alias{.make_off}
\title{QPAD offsets, wrapped by the \code{wt_qpad_offsets} function.}
\usage{
.make_off(spp, x)
}
\arguments{
\item{spp}{species for offset calculation.}

\item{x}{Dataframe out from the \code{.make_x} function.}
}
\description{
Functions to get the offsets.
}
