% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_qpad.R
\name{.make_x}
\alias{.make_x}
\title{Internal functions}
\usage{
.make_x(data, tz = "local", check_xy = TRUE)
}
\description{
Functions to format reports for qpad offset calculation, and then get the offsets. Wrapped by the \code{wt_qpad_offsets} function.
}
\details{
QPAD offsets
}
