% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.wt_api_gr}
\alias{.wt_api_gr}
\title{An internal function to handle generic GET requests to WildTrax API}
\usage{
.wt_api_gr(path, ...)
}
\arguments{
\item{path}{The path to the API}

\item{...}{Argument to pass along into GET query}
}
\description{
Generic function to handle certain GET requests
}
\keyword{internal}
