% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.wt_api_pr}
\alias{.wt_api_pr}
\title{An internal function to handle generic POST requests to WildTrax API}
\usage{
.wt_api_pr(path, ...)
}
\arguments{
\item{path}{The path to the API}

\item{...}{Argument to pass along into POST query}
}
\description{
Generic function to handle certain POST requests
}
