% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{.wt_col_switch}
\alias{.wt_col_switch}
\title{Column switch assignments}
\format{
An object of class \code{list} of length 13.
}
\usage{
.wt_col_switch
}
\description{
Assign correct column names
}
\keyword{internal}
