% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.wt_col_types}
\alias{.wt_col_types}
\title{Column assignments}
\usage{
.wt_col_types(sensor_id = NULL)
}
\arguments{
\item{sensor_id}{Character. Sensor type (e.g. "PC", "ARU"). Determines
type overrides for columns that differ between sensors.}
}
\description{
Assign correct column types for reports, varying by sensor
}
\keyword{internal}
