% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{.wt_col_types}
\alias{.wt_col_types}
\title{Column assignments}
\format{
An object of class \code{list} of length 109.
}
\usage{
.wt_col_types
}
\description{
Assign correct column types for reports
}
\keyword{internal}
