% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wildRtrax.R
\docType{package}
\name{wildRtrax}
\alias{wildRtrax}
\alias{wildRtrax-package}
\title{wildRtrax: Am R package to facilitate full-cycle environmental sensor data work flows.}
\description{
This package provides a set of functions to manage environmental sensor data to and from the WildTrax platform, as well as perform basic data cleaning, analysis, and quality control.
}
\author{
Alex MacPhail \email{agmacpha@ualberta.ca} Marcus Becker \email{mabecker89@ualberta.ca} Elly Knight \email{ecknight@ualberta.ca}
}
