% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wildRtrax.R
\docType{package}
\name{wildRtrax}
\alias{wildRtrax}
\alias{wildRtrax-package}
\title{wildRtrax: A series of advanced user functions for environmental sensor data management and an}
\description{
wildRtrax is an R package that enables environmental sensor data users to create comprehensive work flows for managing and analyzing data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://abbiodiversity.github.io/wildRtrax/}
}

}
\author{
Alex MacPhail \email{agmacpha@ualberta.ca} Marcus Becker \email{mabecker89@ualberta.ca} Elly Knight \email{ecknight@ualberta.ca}
}
