% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wildRtrax.R
\docType{package}
\name{wildRtrax}
\alias{wildRtrax}
\alias{wildRtrax-package}
\title{wildRtrax: A series of advanced user functions for environmental sensor data management, analytics and exploration.}
\description{
This package provides a set of functions to query environmental sensor data from the WildTrax platform, as well as perform basic data cleaning, analysis, and quality control.
}
\author{
Alex MacPhail \email{agmacpha@ualberta.ca} Marcus Becker \email{mabecker89@ualberta.ca}
}
