% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wildtrax.R
\docType{package}
\name{wildtrax}
\alias{wildtrax}
\alias{wildtrax-package}
\title{wildtrax: Data extraction from the Wildtrax platform in R.}
\description{
This package provides functionality to query data from the environmental sensor platform, Wildtrax.
}
\author{
Marcus Becker \email{mabecker@ualberta.ca}

Alex MacPhail \email{agmacpha@ualberta.ca}
}
