% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wildRtrax.R
\docType{package}
\name{wildtrax}
\alias{wildRtrax}
\alias{wildRtrax-package}
\title{wildRtrax: Queries, cleaning, analysis, and quality control of data from the WildTrax database in R}
\description{
This package provides a set of functions for environmental sensor data management and analysis to and from the WildTrax platform. 
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mabecker89/wildRtrax}
  \item Report bugs at \url{https://github.com/mabecker89/wildRtrax/issues}
}

}
\author{
Marcus Becker \email{mabecker@ualberta.ca}

Alex MacPhail \email{agmacpha@ualberta.ca}
}
