% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifier-functions.R
\name{wt_additional_species}
\alias{wt_additional_species}
\title{Find new species}
\usage{
wt_additional_species(
  data,
  remove_species = TRUE,
  threshold = 50,
  resolution = "task"
)
}
\arguments{
\item{data}{Output from the \code{wt_download_report()} function when you request the \code{main} and \code{birdnet} reports}

\item{remove_species}{Logical; indicates whether species that are not allowed in the WildTrax project should be removed from the BirdNET report}

\item{threshold}{Numeric; the desired score threshold}

\item{resolution}{Character; either "recording" to identify any new species for each recording or "location" to identify new species for each location}
}
\value{
A tibble with the same fields as the \code{birdnet} report with the highest scoring detection for each new species detection in each recording.
}
\description{
Check for species reported by BirdNET that the human listeners did not detect in our project.
}
\examples{
\dontrun{
data <- wt_download_report(project_id = 1144, sensor_id = "ARU",
reports = c("main", "birdnet"), weather_cols = FALSE)

new <- wt_additional_species(data, remove_species = TRUE,
threshold = 80, resolution="location")
}

}
