% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acoustic-pre-processing.R
\name{wt_audio_scanner}
\alias{wt_audio_scanner}
\title{Scan acoustic data to a standard format}
\usage{
wt_audio_scanner(path, file_type, extra_cols = F, tz = "")
}
\arguments{
\item{path}{Character; The path to the directory with audio files you wish to scan. Can be done recursively.}

\item{file_type}{Character; Takes one of four values: wav, wac, flac or all. Use "all" if your directory contains many types of files.}

\item{extra_cols}{Boolean; Default set to FALSE for speed. If TRUE, returns additional columns for file duration, sample rate and number of channels.}

\item{tz}{Character; Forces a timezone to each of the recording files; if the time falls into a daylight savings time break, \code{wt_audio_scanner} will assume the next valid time. Use \code{OlsonNames()} to get a list of valid names.}
}
\value{
A tibble with a summary of your audio files. Includes file path, file name, file size, date, time, location name, sample rate, length (seconds) and number of channels to be used as filters for other uses
}
\description{
Scans directories of audio data and returns the standard naming conventions
}
\examples{
wt_audio_scanner(path = ".", file_type = "wav", extra_cols = T, tz = "US/Mountain")


}
