% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acoustic-pre-processing.R
\name{wt_audio_scanner}
\alias{wt_audio_scanner}
\title{Scan acoustic data to a standard format}
\usage{
wt_audio_scanner(path, file_type, extra_cols = F)
}
\arguments{
\item{path}{Character; The path to the directory with audio files you wish to scan. Can be done recursively.}

\item{file_type}{Character; Takes one of four values: wav, wac, flac or all. Use "all" if your directory contains many types of files.}

\item{extra_cols}{Boolean; Default set to FALSE for speed. If TRUE, returns additional columns for file duration, sample rate and number of channels.}
}
\value{
A tibble with a summary of your audio files.
}
\description{
Scans directories of audio data and returns the standard naming conventions
}
\examples{
\dontrun{
wt_audio_scanner(path = ".", file_type = "wav", extra_cols = T)
}

}
