% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wt_audio_scanner.R
\name{wt_audio_scanner}
\alias{wt_audio_scanner}
\title{Scans directories of audio data and returns the filepath, filename, file size, date, time, station key, sample rate, length /(s)/ and number of channels to be used as filters for wt_aru_assign or other uses}
\usage{
wt_audio_scanner(path, file_type)
}
\arguments{
\item{path}{The path to the directory with audio files you wish to scan (character)}

\item{file_type}{Can be either wac or wav file types (character)}
}
\value{
A dataframe with a summary of your audio files
}
\description{
Scans directories of audio data and returns the filepath, filename, file size, date, time, station key, sample rate, length /(s)/ and number of channels to be used as filters for wt_aru_assign or other uses
}
\examples{
\dontrun{
df <- wt_audio_scanner(path = "my_path", file_type = "\\\\.wav$|\\\\.wac$")
}

}
