% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wt_audio_scanner.R
\name{wt_audio_scanner}
\alias{wt_audio_scanner}
\title{Scan and extract metadata from audio data}
\usage{
wt_audio_scanner(path, file_type)
}
\arguments{
\item{path}{Character; The path to the directory with audio files you wish to scan. Can be done recursively.}

\item{file_type}{Character; Takes one of three values: wav, wac, or both. Use "both" if your directory contains both types of files.}
}
\value{
A dataframe with a summary of your audio files
}
\description{
Scans directories of audio data and returns the file path, file name, file size, date, time, location name,
sample rate, length (seconds) and number of channels to be used as filters for other uses
}
\examples{
\dontrun{
df <- wt_audio_scanner(path = "C:/Users/me/path/to/audio/files", file_type = "both")
}

}
