% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acoustic-pre-processing.R
\name{wt_audio_scanner}
\alias{wt_audio_scanner}
\title{Scan acoustic data to a standard format}
\usage{
wt_audio_scanner(path, file_type, extra_cols = F, tz = "")
}
\arguments{
\item{path}{Character; The path to the directory with audio files you wish to scan. Can be done recursively.}

\item{file_type}{Character; Takes one of four values: wav, wac, flac or all. Use "all" if your directory contains many types of files.}

\item{extra_cols}{Boolean; Default set to FALSE for speed. If TRUE, returns additional columns for file duration, sample rate and number of channels.}

\item{tz}{Character; Forces a timezone to each of the recording files; if the time falls into a daylight savings time break, \code{wt_audio_scanner} will assume the next valid time. Use \code{OlsonNames()} to get a list of valid names.}
}
\value{
A tibble with a summary of your audio files.
\itemize{
\item file_path; Character - Full file path of the audio
\item size_Mb; Numeric - Size of the file in MB
\item file_name; Character - Basename of the file
\item location; Character - Location name
\item recording_date_time; Dttm - Recording date time of the file; use with "tz" argument to set as needed
\item file_type; Character - wac, wav or flac
\item julian; Numeric - Julian date of the audio
\item year; Numeric - Year of the audio
\item gps_enabled; Boolean - If TRUE, audio is GPS enabled
\item time_index; Numeric - Numeric value describing the order of the recording per day. E.g. time_index = 1 is the first recording of the day
\item length_seconds; Numeric - Length of the audio file in seconds
\item sample rate; Numeric - Sample rate of the audio file in Hertz
\item n_channels; Numeric - Number of channels of the audio file
}
}
\description{
Scans directories of audio data and returns the standard naming conventions
}
\examples{
\dontrun{
wt_audio_scanner(path = ".", file_type = "wav", extra_cols = T, tz = "US/Mountain")
}

}
