% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acoustic-pre-processing.R
\name{wt_chop}
\alias{wt_chop}
\title{Segment large audio files}
\usage{
wt_chop(input = NULL, segment_length = NULL, output_folder = NULL)
}
\arguments{
\item{input}{A data frame or tibble containing information about audio files}

\item{segment_length}{Numeric; Segment length in seconds. Modulo recording will be exported should there be any trailing time left depending on the segment length used}

\item{output_folder}{Character; output path to where the segments will be stored}
}
\value{
No return value, called for file-writing side effects.
}
\description{
"Chops" up wav files into many smaller files of a desired duration and writes them to an output folder.
}
\examples{
\dontrun{
wt_chop(input = my_files, segment_length = 60, output_folder = "output_folder")
}

}
