% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifier-functions.R
\name{wt_classifier_threshold}
\alias{wt_classifier_threshold}
\title{Identify optimal threshold}
\usage{
wt_classifier_threshold(data)
}
\arguments{
\item{data}{Tibble output from the \code{wt_evaluate_classifier()} function.}
}
\value{
A single numeric value
}
\description{
Retrieves the score threshold that maximizes F-score, which is a trade-off between precision and recall.
}
\examples{
\dontrun{
data <- wt_download_report(project_id = 1144, sensor_id = "ARU",
reports = c("main", "ai"))

eval <- wt_evaluate_classifier(data, resolution = "recording",
remove_species = TRUE, thresholds = c(10, 99))

threshold_use <- wt_classifier_threshold(eval) |> print()
}

}
