% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wt_con.R
\name{wt_con}
\alias{wt_con}
\title{Establish Connection to Wildtrax PostgreSQL Database}
\usage{
wt_con(username, password)
}
\arguments{
\item{username}{A character string indicating the username for access the Wildtrax database}

\item{password}{A character string indicating the password associated with the supplied username}
}
\value{
This function returns an S4 object of class PostgreSQLConnection
}
\description{
Establish Connection to Wildtrax PostgreSQL Database
}
\author{
Marcus Becker

Alex MacPhail
}
