% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wt_download_clips.R
\name{wt_download_clips}
\alias{wt_download_clips}
\title{Download clips of WildTrax annotations.}
\usage{
wt_download_clips(data, filepath, object = "soundfile", quiet = FALSE)
}
\arguments{
\item{data}{WildTrax tag report from the \code{wt_download_report} function.}

\item{filepath}{Character; folder to download objects into.}

\item{object}{Character; the format requested for the clips ("soundfile", or "spectrogram").}

\item{quiet}{Logical; suppress download details for each object.}
}
\value{
A dataframe identical to input but including ID and filepath for each downloaded object.
}
\description{
This function downloads annotations from WildTrax either as sound files or spectrograms. Using this function requires obtaining the urls for the objects using the \code{wt_download_report} function.
}
\examples{
\dontrun{
dat.tag <- wt_download_clips(dat, "data/clips", "soundfile", quiet=TRUE)
}
}
