% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{wt_download_media}
\alias{wt_download_media}
\title{Download acoustic media}
\usage{
wt_download_media(
  input_data,
  output,
  type = c("recording", "tag_clip_audio", "tag_clip_spectrogram")
)
}
\arguments{
\item{output}{The output folder}

\item{type}{Either recording, tag_clip_spectrogram or tag_clip_audio}

\item{input}{The report data}
}
\value{
An organized folder of media. Assigning wt_download_tags to an object will return the table form of the data with the functions returning the after effects in the output directory
}
\description{
Download acoustic media in batch
}
\examples{
\dontrun{
dat.report <- wt_download_report() |>
wt_download_media(output = "my/output/folder")
}

}
