% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{wt_download_media}
\alias{wt_download_media}
\title{Download media}
\usage{
wt_download_media(
  input,
  output,
  type = c("recording", "image", "tag_clip_audio", "tag_clip_spectrogram")
)
}
\arguments{
\item{input}{The report data}

\item{output}{The output folder}

\item{type}{Either recording, image, tag_clip_spectrogram, tag_clip_audio}
}
\value{
An organized folder of media. Assigning wt_download_tags to an object will return the table form of the data with the functions returning the after effects in the output directory
}
\description{
Download acoustic and image media in batch. Includes the download of tag clips and spectrograms for the ARU sensor.
}
\examples{
\dontrun{
dat.report <- wt_download_report() |>
wt_download_media(output = "my/output/folder", type = "recording")
}

}
