% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{wt_download_report}
\alias{wt_download_report}
\title{Get download summary}
\usage{
wt_download_report(
  project_id,
  sensor_id,
  cols_def = FALSE,
  weather_cols = TRUE
)
}
\arguments{
\item{project_id}{Numeric; the project ID number that you would like to download data for. Use \verb{wt_get_download_summary() to retrieve these IDs.}}

\item{sensor_id}{Can either be "ARU" or "CAM"}

\item{cols_def}{Logical; Do you want to include the column definitions? Defaults to FALSE}

\item{weather_cols}{Logical; Do you want to include weather information for your stations? Defaults to TRUE}
}
\value{
If cols_def is TRUE, a list is returned with three elements: the data, english, and french column definitions; if FALSE, the data is returned as a dataframe.
}
\description{
Obtain a table listing projects that the user is able to download data for
}
\examples{
\dontrun{
# Authenticate first:
wt_auth()
wt_get_download_summary(sensor_id = "ARU")
}

}
