% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{wt_download_report}
\alias{wt_download_report}
\title{Download Reports}
\usage{
wt_download_report(project_id, sensor_id, reports, weather_cols = TRUE)
}
\arguments{
\item{project_id}{Numeric; the project ID number that you would like to download data for. Use \code{wt_get_download_summary()} to retrieve these IDs.}

\item{sensor_id}{Character; Can either be "ARU", "CAM", or "PC".}

\item{weather_cols}{Logical; Do you want to include weather information for your stations? Defaults to TRUE.}

\item{report}{Character; The report type to be returned. Multiple values are accepted as a concatenated string.}
}
\value{
If multiple report types are requested, a list object is returned; if only one, a dataframe.
}
\description{
Download ARU, Camera, or Point Count data from a project
}
\details{
Valid values for argument \code{report} when \code{sensor_id} = "CAM" currently are:
\itemize{
\item main
\item project
\item location
\item image
\item tag
\item megadetector
\item megaclassifier
\item definitions
}

Valid values for argument \code{report} when \code{sensor_id} = "ARU" currently are:
\itemize{
\item main
\item project
\item location
\item tag
\item birdnet
\item definitions
}

Valid values for argument \code{report} when \code{sensor_id} = "PC" currently are:
\itemize{
\item main
\item project
\item location
\item point count
\item definitions
}
}
\examples{
\dontrun{
# Authenticate first:
wt_auth()
a_camera_project <- wt_download_report(project_id = 397, sensor_id = "CAM", reports = c("tag", "image"), weather_cols = TRUE)

an_aru_project <- wt_download_report(project_id = 47, sensor_id = "ARU", reports = c("main", "birdnet"), weather_cols = TRUE)
}

}
