% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifier-functions.R
\name{wt_evaluate_classifier}
\alias{wt_evaluate_classifier}
\title{Evaluate a classifier}
\usage{
wt_evaluate_classifier(
  data,
  resolution = "recording",
  remove_species = TRUE,
  species = NULL,
  thresholds = c(10, 99)
)
}
\arguments{
\item{data}{Output from the \code{wt_download_report()} function when you request the \code{main} and \code{birdnet} reports}

\item{resolution}{Character; either "recording" to summarize at the entire recording level or "minute" to summarize the minute level if the \code{task_method} is "1SPM", or "task"}

\item{remove_species}{Logical; indicates whether species that are not allowed in the WildTrax project should be removed from the BirdNET report}

\item{species}{Character; optional subset of species to calculate metrics for (e.g., species = c("OVEN", "OSFL", "BOCH"))}

\item{thresholds}{Numeric; start and end of sequence of score thresholds at which to calculate performance metrics}
}
\value{
A tibble containing columns for precision, recall, and F-score for each of the requested thresholds.
}
\description{
Calculates precision, recall, and F-score of BirdNET for a requested sequence of thresholds. You can request the metrics at the minute level for recordings that are processed with the species per minute method (1SPM). You can also exclude species that are not allowed in the project from the BirdNET results before evaluation.
}
\examples{
\dontrun{
data <- wt_download_report(project_id = 1144, sensor_id = "ARU",
reports = c("main", "birdnet"), weather_cols = FALSE)

eval <- wt_evaluate_classifier(data, resolution = "recording",
remove_species = TRUE, thresholds = c(10, 99))
}

}
