% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{wt_get_download_summary}
\alias{wt_get_download_summary}
\title{Get a download summary from WildTrax}
\usage{
wt_get_download_summary(sensor_id)
}
\arguments{
\item{sensor_id}{Can be one of "ARU", "CAM", or "PC"}
}
\value{
A dataframe listing the projects that the user can download data for, including: project name, id, year, number of tasks, a geographic bounding box and project status.
}
\description{
Obtain a table listing projects that the user is able to download data for
}
\examples{
\dontrun{
# Authenticate first:
wt_auth()
wt_get_download_summary(sensor_id = "ARU")
}

}
