% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience-functions.R
\name{wt_get_exif}
\alias{wt_get_exif}
\title{Get EXIF metadata from images}
\usage{
wt_get_exif(data)
}
\arguments{
\item{data}{\code{wt_download_report(reports = c(image_report))} object containing}
}
\value{
A dataframe with the EXIF metadata for each image
}
\description{
This function gets all relevant EXIF metadata from images in Projects
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
\dontrun{

dat <- wt_download_report(reports = c("image_report"))
exif.data <- wt_get_exif(dat)
}
}
