% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{wt_get_locations}
\alias{wt_get_locations}
\title{Get locations from a WildTrax Organization}
\usage{
wt_get_locations(organization)
}
\arguments{
\item{organization}{Either the short letter or numeric digit representing the Organization}
}
\value{
A data frame listing an Organizations' locations
}
\description{
Obtain a table listing locations emulating the Locations tab in a WildTrax Organization
}
\examples{
\dontrun{
# Authenticate first:
wt_auth()
wt_get_locations(organization = 'ABMI')
}

}
