% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{wt_get_project_species}
\alias{wt_get_project_species}
\title{Get the WildTrax species list for a specific project}
\usage{
wt_get_project_species(project)
}
\arguments{
\item{project}{The project id associated with the desired species list}
}
\value{
A tibble of the WildTrax species list for a specific project
}
\description{
Request for the WildTrax species list for a project
}
\examples{
 \dontrun{
 my_project_species <- wt_get_project_species()
 }
}
