% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{wt_get_project_species}
\alias{wt_get_project_species}
\title{Get project-species from WildTrax project}
\usage{
wt_get_project_species(project_id)
}
\arguments{
\item{project_id}{The project_id of the WildTrax project}
}
\value{
A data frame listing an Organizations' locations
}
\description{
Obtain a table listing species added to a specific project in WildTrax
}
\examples{
\dontrun{
# Authenticate first:
wt_auth()
my_project <- wt_get_download_summary(sensor_id = 'ARU') |>
filter(grepl('Ecosystem Health 2023',project)) |>
pull(project_id)
wt_get_project_species(project_id = my_project)
}

}
