% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{wt_get_species}
\alias{wt_get_species}
\title{Get the WildTrax species table}
\usage{
wt_get_species()
}
\value{
A tibble of the species table called 'wt_spp_table'
}
\description{
Request for the WildTrax species table
}
\examples{
 \dontrun{
 wt_species <- wt_get_species()
 }
}
