% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{wt_get_sync}
\alias{wt_get_sync}
\title{Get column headers and data from WildTrax Sync APIs}
\usage{
wt_get_sync(
  api,
  option = c("columns", "data"),
  project = NULL,
  organization = NULL
)
}
\arguments{
\item{api}{A string specifying the API to query. Must be one of:
\itemize{
\item \code{"download-location-by-org-id"}
\item \code{"download-visits-by-org-id"}
\item \code{"download-equipment-by-org-id"}
\item \code{"download-location-equipment-by-org-id"}
\item \code{"download-location"}
\item \code{"download-tasks-by-project-id"}
\item \code{"download-tags-by-project-id"}
\item \code{"download-camera-tags-by-project-id"}
\item \code{"download-point-count-by-project-id"}
}}

\item{option}{An option}
}
\value{
A tibble with column headers for the specified API call.
}
\description{
Fetch column headers and data for sync APIs in WildTrax. You must specify at least one of \code{project} or \code{organization} depending on the API and at what level of the system you're looking for data for
}
\examples{
\dontrun{
# Authenticate first:
wt_auth()

# Fetch column headers by organization
wt_get_sync_columns("download-location-by-org-id")

# Fetch column headers by project
wt_get_sync_columns("download-point-count-by-project-id")
}

}
