% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{wt_get_sync}
\alias{wt_get_sync}
\title{Get column headers and data from WildTrax Sync APIs}
\usage{
wt_get_sync(
  api,
  option = c("columns", "data"),
  project = NULL,
  organization = NULL
)
}
\arguments{
\item{api}{A string specifying the API to query. Must be one of:
\itemize{
\item \code{"organization_locations"} - Fetch location summaries for an organization (\code{get-location-summary}).
\item \code{"organization_visits"} - Fetch visit summaries for an organization (\code{get-location-visit-summary}).
\item \code{"organization_equipment"} - Fetch equipment details for an organization (\code{get-equipment}).
\item \code{"recording_summary"} - Fetch recording summaries for an organization (\code{get-recording-summary}).
\item \code{"image_deployment_summary"} - Fetch image deployment summaries for an organization (\code{get-image-deployment-summary}).
\item \code{"project_locations"} - Fetch location details for a project (\code{download-location}).
\item \code{"project_tasks"} - Fetch task summaries for a project (\code{download-tasks-by-project-id}).
\item \code{"project_tags"} - Fetch tag summaries for a project (\code{download-tags-by-project-id}).
\item \code{"project_camera_tags"} - Fetch camera tag summaries for a project (\code{download-camera-tags-by-project-id}).
\item \code{"project_point_counts"} - Fetch point count summaries for a project (\code{download-point-count-by-project-id}).
\item \code{"project_species"} - Fetch species details for a project (\code{get-project-species-details}).
}}

\item{option}{An option}
}
\value{
A tibble with column headers for the specified API call.
}
\description{
Fetch column headers and data for sync APIs in WildTrax. You must specify at least one of \code{project} or \code{organization} depending on the API and at what level of the system you're looking for data for
}
\examples{
\dontrun{
# Authenticate first:
wt_auth()

# Fetch column headers by organization
wt_get_sync_columns("download-location-by-org-id")

# Fetch column headers by project
wt_get_sync_columns("download-point-count-by-project-id")
}

}
