% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{wt_get_sync}
\alias{wt_get_sync}
\title{Get data from WildTrax syncs and downloads}
\usage{
wt_get_sync(api, project = NULL, organization = NULL)
}
\arguments{
\item{api}{A string specifying the API to query. Must be one of:
\itemize{
\item \code{"organization_locations"}
\item \code{"organization_visits"}
\item \verb{"organization_equipment}"
\item \verb{"organization_deployments}"
\item \verb{"organization_recordings}"
\item \code{"project_locations"}
\item \code{"project_species"}
\item \code{"project_aru_tasks"}
\item \code{"project_aru_tags"}
\item \verb{"project_image_sets}"
\item \code{"project_image_metadata"}
\item \code{"project_camera_tags"}
\item \code{"project_point_counts"}
}}

\item{project}{The project id}

\item{organization}{The organization id}
}
\value{
A tibble with column headers for the specified API call.
}
\description{
Fetch data for syncs and downloads in WildTrax. You must specify at least one of \code{project} or \code{organization} depending on the API and at what level of the system you're looking for data for. Use \code{wt_download_report()}.
}
\examples{
\dontrun{
# Authenticate first:
wt_auth()

# Fetch column headers by organization
wt_get_sync("organization_locations", organization = "ABMI")

# Fetch column headers by project
wt_get_sync("project_locations", project = 620)
}

}
