% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{wt_get_sync_columns}
\alias{wt_get_sync_columns}
\title{Get column headers from WildTrax Sync APIs}
\usage{
wt_get_sync_columns(api, project = NULL, organization = NULL)
}
\arguments{
\item{api}{A string specifying the API to query. Must be one of:
\itemize{
\item \code{"download-location-by-org-id"}
\item \code{"download-point-count-by-project-id"}
}}

\item{project}{The project ID to query. (Optional if \code{organization} is provided.)}

\item{organization}{The organization ID to query. (Optional if \code{project} is provided.)}
}
\value{
A tibble with column headers for the specified API call.
}
\description{
Fetch column headers for a sync APIs in WildTrax. You must specify at least one of \code{project} or \code{organization}.
}
\examples{
\dontrun{
# Authenticate first:
wt_auth()

# Fetch column headers by organization
wt_get_sync_columns("download-location-by-org-id", organization = 5)

# Fetch column headers by project
wt_get_sync_columns("download-point-count-by-project-id", project = 3355)
}

}
