% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acoustic-pre-processing.R
\name{wt_guano_tags}
\alias{wt_guano_tags}
\title{Convert GUANO embeded metadata to tags and metadata output for a Project}
\usage{
wt_guano_tags(path, output = NULL, output_file = NULL)
}
\arguments{
\item{path}{Character; The path to the input csv}

\item{output}{Character; Path where the output file will be stored}

\item{output_file}{Character; Path of the output file}
}
\value{
A csv formatted as a WildTrax tag template
}
\description{
\code{wt_guano_tags} Takes the embeded classifier output and converts them into a WildTrax tag template for upload
}
\examples{
\dontrun{
wt_guano_tags(path = my_audio_file.csv, output = NULL, output_file = NULL)
}

}
