% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acoustic-pre-processing.R
\name{wt_kaleidoscope_tags}
\alias{wt_kaleidoscope_tags}
\title{Upload hits from a Wildlife Acoustics classifier}
\usage{
wt_kaleidoscope_tags(input, output, tz, freq_bump = T)
}
\arguments{
\item{input}{Character; The path to the input csv}

\item{output}{Character; Path where the output file will be stored}

\item{tz}{Character; Assigns a timezone to the recording files. Use \code{OlsonNames()} to get a list of valid names.}

\item{freq_bump}{Boolean; Set to TRUE to add a buffer to the frequency values exported from Kaleidoscope. Helpful for getting more context around a signal in species verification}
}
\value{
A csv formatted as a WildTrax tag template
}
\description{
\code{wt_kaleidoscope_tags} Takes the classifier output from Wildlife Acoustics Kaleidoscope and converts them into a WildTrax tag template for upload
}
\section{\code{wt_kaleidoscope_tags}}{
NA
}

\examples{
\dontrun{
wt_kaleidoscope_tags(input = input.csv, output = tags.csv, tz = "", freq_bump = T)
}

}
