% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience-functions.R
\name{wt_location_distances}
\alias{wt_location_distances}
\title{General convenience functions}
\usage{
wt_location_distances(input_from_tibble = NULL, input_from_file = NULL)
}
\arguments{
\item{input_from_tibble}{Use a tibble constructed with a distinct list of location names, latitude and longitude}

\item{input_from_file}{Use a file downloaded from either an organization or project}
}
\value{
A three-column tibble with the distances between each location
}
\description{
Takes input lats and longs and computes the distances between each set of points
}
\examples{
\dontrun{
df <- wt_location_distances(input = my_location_tibble, input_from_file)
}

}
