% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{wt_location_photos}
\alias{wt_location_photos}
\title{Batch download location photos}
\usage{
wt_location_photos(organization, output = NULL)
}
\arguments{
\item{organization}{Character; The Organization acronym from which photos are being downloaded}

\item{output}{Character; Directory where location photos would be stored on your machine}

\item{download_all}{Logical; If \code{TRUE}, defaults to downloading all your location photos}
}
\value{
A folder of photos or an object containing the download information for said photos
}
\description{
Download location photos from an Organization
}
\examples{
\dontrun{
# Authenticate first:
wt_auth()
wt_location_photos(organization = 'ABMI', output = NULL)
}

}
