% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience-functions.R
\name{wt_make_wide}
\alias{wt_make_wide}
\title{Convert to a wide survey by species dataframe}
\usage{
wt_make_wide(data, sound = "all")
}
\arguments{
\item{data}{WildTrax main report or tag report from the \code{wt_download_report()} function.}

\item{sound}{Character; vocalization type(s) to retain ("all", "Song", "Call", "Non-vocal"). Can be used to remove certain types of detections. Defaults to "all" (i.e., no filtering).}
}
\value{
A dataframe identical to input with observations of the specified groups removed.
}
\description{
This function converts a long-formatted report into a wide survey by species dataframe of abundance values.
}
\examples{
\dontrun{
dat.tidy <- wt_tidy_species(dat)
dat.tmtt <- wt_replace_tmtt(dat.tidy)
dat.wide <- wt_make_wide(dat.tmtt, sound="all")
}
}
