% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience-functions.R
\name{wt_replace_tmtt}
\alias{wt_replace_tmtt}
\title{Replace 'TMTT' abundance with model-predicted values}
\usage{
wt_replace_tmtt(data, calc = "round")
}
\arguments{
\item{data}{Dataframe of WildTrax observations, for example the summary report.}

\item{calc}{Character; method to convert model predictions to integer ("round", "ceiling", or "floor"). See \code{?round()} for details.}
}
\value{
A dataframe identical to input with 'TMTT' entries in the abundance column replaced by integer values.
}
\description{
This function uses a lookup table of model-predicted values to replace 'TMTT' entries in listener-processed ARU data from WildTrax. The model-predicted values were produced using estimated abundances for 'TMTT' entries in mixed effects model with a Poisson distribution and random effects for species and observer.
}
\examples{
\dontrun{
dat.tmtt <- wt_replace_tmtt(data, calc="round")
}
}
