% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting.R
\name{wt_report}
\alias{wt_report}
\title{Build automatic reports from analyzed data in wildRtrax}
\usage{
wt_report(data, what = c(NULL, "conclusion"), output_filename = "report")
}
\arguments{
\item{data}{Input data from wt_download_report}

\item{what}{NULL, conclusion}

\item{output_filename}{Character; name of the output file name}
}
\value{
A document with the report
}
\description{
Build reports automatically using custom settings
}
\section{\code{wt_report} details}{


Build customized reports from WildTrax data
}

\examples{
\dontrun{
wt_report(data = my_data, what = c(NULL, "conclusion"))
}

}
