% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wt_run_ap.R
\name{wt_run_ap}
\alias{wt_run_ap}
\title{Generate acoustic indices for a folder of audio files with QUT Ecoacoustics \strong{A}nalysis \strong{P}rograms (AP.exe) software}
\usage{
wt_run_ap(
  x = NULL,
  fp_col = file_path,
  audio_dir = NULL,
  output_dir,
  path_to_ap = "C:\\\\AP\\\\AnalysisPrograms.exe"
)
}
\arguments{
\item{x}{(optional) A data frame or tibble; must contain the direct audio file path, name and size. Use output from \code{wt_audio_scanner}.}

\item{fp_col}{If x is supplied, the column containing the audio file paths. Defaults to file_path.}

\item{audio_dir}{(optional) Character; path to directory storing audio files.}

\item{output_dir}{Character; path to directory where you want outputs to be stored.}

\item{path_to_ap}{Character; file path to the AnalysisPrograms software package. Defaults to "C:\\AP\\AnalysisPrograms.exe".}
}
\value{
Output will return to the specific root directory
}
\description{
See \url{https://github.com/QutEcoacoustics/audio-analysis} for information about usage and installation of the AP software. Note that this function relies on having this software installed locally.

This function will batch calculate summary and spectral acoustic indices for a folder of audio files using the Towsey.Acoustic configuration (yml) file from the AP software. You can use the output from \code{wt_audio_scanner} in the function, or define a local folder with audio files directly.
}
\examples{
\dontrun{
}

}
