% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linking-media-to-wildtrax.R
\name{wt_songscope_tags}
\alias{wt_songscope_tags}
\title{\subsection{Sonic hits}{
}}
\usage{
wt_songscope_tags(input, output, species_code, vocalization_type, task_length)
}
\arguments{
\item{input}{Character; The path to the input csv}

\item{output}{Character; Path where the output file will be stored}

\item{species_code}{Character;}

\item{vocalization_type}{Character;}

\item{task_length}{Numeric;}
}
\value{
A csv formatted as a WildTrax tag template
}
\description{
Takes the classifier output from Wildlife Acoustics Songscope and converts them into a WildTrax tag template for upload
}
\examples{
\dontrun{
wt_songscope_tags(input = input.csv, output = tags.csv, species_code, vocalization_type = "Call", task_length = 180)
}

}
