% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze-data.R
\name{wt_summarise_cam}
\alias{wt_summarise_cam}
\title{Set of analysis functions}
\usage{
wt_summarise_cam(
  detect_data,
  raw_data,
  time_interval = "day",
  variable = "detections",
  output_format = "wide",
  species_col = species_common_name,
  effort_data = NULL,
  project_col = NULL,
  station_col = NULL,
  start_col = NULL,
  end_col = NULL
)
}
\arguments{
\item{detect_data}{Detection data generated from \code{wt_ind_det()}.}

\item{raw_data}{The raw camera tag data, which is used to infer the effort (i.e. date ranges of operation) for each camera. Optionally, can supply effort_data directly instead.}

\item{time_interval}{Character; Can be either "full", "month", "week", or "day".}

\item{variable}{Character; Can be either "presence", "detections", "counts", or "all" (if you want all three).}

\item{output_format}{Character; The format of the dataframe returned to you. Can be either "wide" or "long".}

\item{species_col}{Defaults to \code{common_name}. The column referring to species. Use to switch between common and scientific names of species, if you have both.}

\item{effort_data}{Optionally supply your own effort data.}

\item{project_col}{The column referring to project in your effort data.}

\item{station_col}{The column referring to each individual camera station/location in your effort data.}

\item{start_col}{The column indicating the start date of the camera location}

\item{end_col}{The column indicating the end date of the camera location}
}
\value{
A dataframe summarising your camera data by location, time interval, and species.
}
\description{
This function takes your independent detection data and summarises it by location, specified time interval, and species.
}
\section{Summarise camera data by location, time interval, and species.}{
NA
}

\examples{
\dontrun{
summary <- wt_summarise_cam(
x, y, time_interval = "day", variable = "detections", output_format = "wide"
)
}

}
