% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience-functions.R
\name{wt_tidy_species}
\alias{wt_tidy_species}
\title{Filter the species list to the groups of interest from wt_get_species}
\usage{
wt_tidy_species(
  data,
  remove = c("mammal", "amphibian", "abiotic", "insect", "unknown"),
  zerofill = TRUE
)
}
\arguments{
\item{data}{WildTrax main report or tag report from the \code{wt_download_report} function.}

\item{remove}{Character; groups to filter from the report ("mammal", "bird", "amphibian", "abiotic", "insect", "unknown"). Defaults to retaining bird group only.}

\item{zerofill}{Logical; indicates if zerofilling should be completed. If TRUE, unique surveys with no observations after filtering are added to the dataset with "NONE" as the value for species_code and/or species_common_name. If FALSE, only surveys with observations of the retained groups are returned. Default is TRUE.}
}
\value{
A dataframe identical to input with observations of the specified groups removed.
}
\description{
This function filters the species provided in WildTrax reports to only the groups of interest. The groups available for filtering are mammal, bird, amphibian, abiotic, insect, and unknown. Zero-filling functionality is available to ensure all surveys are retained in the dataset if no observations of the group of interest are available.
}
\examples{
\dontrun{
dat.tmtt <- wt_tidy_species(dat, remove=c("mammal", "amphibian", "abiotic", "insect", "unknown"), zerofill = TRUE)
}
}
