% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acoustic-pre-processing.R
\name{wt_wac_info}
\alias{wt_wac_info}
\title{Extract relevant metadata from a wac file}
\usage{
wt_wac_info(path)
}
\arguments{
\item{path}{Character; The wac file path}
}
\value{
a list with relevant information
}
\description{
Scrape relevant information from wac (Wildlife Acoustics) file
}
